update for edge tpu

+++ include/tensorflow/lite/micro/memory_planner/linear_memory_planner.h
@@ -31,16 +31,16 @@ class LinearMemoryPlanner : public MemoryPlanner {
   TfLiteStatus AddBuffer(tflite::ErrorReporter* error_reporter, int size,
                          int first_time_used, int last_time_used) override;
 
-  size_t GetMaximumMemorySize() override;
+  int GetMaximumMemorySize() override;
   int GetBufferCount() override;
   TfLiteStatus GetOffsetForBuffer(tflite::ErrorReporter* error_reporter,
                                   int buffer_index, int* offset) override;
 
  private:
   static constexpr int kMaxBufferCount = 1024;
-  size_t buffer_offsets_[kMaxBufferCount];
+  int buffer_offsets_[kMaxBufferCount];
   int current_buffer_count_;
-  size_t next_free_offset_;
+  int next_free_offset_;
 
   TF_LITE_REMOVE_VIRTUAL_DELETE
 };
